requirements.txt
Django==5.1.1
djangorestframework==3.15.2
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.1
